# factory/factory-service/openapi/factory-openapi.yaml
openapi: 3.1.0
info:
  title: Factory Service API
  description: REST API for the Factory Service — autonomous Kotlin/Spring Boot backend
  version: 0.0.1
servers:
- url: http://localhost:18141
  description: Generated server url
...
paths:
  /api/factory/workstreams/{workstreamId}:
    get:
      tags:
      - workstreams
      summary: Read one workstream registry entry of the caller's trusted workstream.
      operationId: getWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
    put:
      tags:
      - workstreams
      summary: Save a new revision of a workstream registry entry.
      operationId: updateWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      - name: If-Match
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkstreamRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
...
  /api/factory/workstreams:
    get:
      tags:
      - workstreams
      summary: List workstreams for the caller's tenant.
      operationId: listWorkstream
      parameters:
      - name: namespaceId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: {}
    post:
      tags:
      - workstreams
      summary: Create a workstream entry.
      operationId: createWorkstream
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkstreamRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
  /api/factory/workstreams/{workstreamId}/controller-case:
    get:
      tags:
      - workstreams
      summary: Read the active controller case of the caller's trusted workstream.
      operationId: activeControllerCaseWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ControllerCaseResponse"
    post:
      tags:
      - workstreams
      summary: Start the first controller case of the caller's trusted workstream.
      operationId: startControllerCaseWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartControllerCaseRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ControllerCaseResponse"
  /api/factory/workstreams/{workstreamId}/controller-case/compact:
    post:
      tags:
      - workstreams
      summary: Compact / renew the controller case of the caller's trusted workstream.
      operationId: compactControllerCaseWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompactControllerCaseRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ControllerCaseResponse"
...
  /api/factory/workstreams/{workstreamId}/projection:
    get:
      tags:
      - workstreams
      summary: Read-only aggregated projection of the caller's trusted workstream.
      operationId: projectionWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      - name: namespaceId
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkstreamProjectionResponse"
  /api/factory/workstreams/{workstreamId}/controller-case/history:
    get:
      tags:
      - workstreams
      summary: List the controller case history (active + archived) of the caller's
        trusted workstream.
      operationId: controllerCaseHistoryWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ControllerCaseHistoryResponse"
  /api/factory/workstreams/{workstreamId}/controller-case/context:
    get:
      tags:
      - workstreams
      summary: Preview the bounded controller resumption context package of the caller's
        trusted workstream.
      operationId: controllerCaseContextWorkstream
      parameters:
      - name: workstreamId
        in: path
        required: true
        schema:
          type: string
      - name: namespaceId
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ControllerResumptionPackage"
...
components:
  schemas:
    UpdateWorkstreamRequest:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        status:
          type: string
        namespaceId:
          type: string
        controllerAgentRef:
          type: string
        allowedWorkflowTypes:
          type: array
          items:
            type: string
        governancePolicyRef:
          type: string
        expectedRevision:
          type: integer
          format: int32
...
    CreateWorkstreamRequest:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        title:
          type: string
        status:
          type: string
        namespaceId:
          type: string
        controllerAgentRef:
          type: string
        allowedWorkflowTypes:
          type: array
          items:
            type: string
        governancePolicyRef:
          type: string
    StartControllerCaseRequest:
      type: object
      properties:
        caseId:
          type: string
    ControllerCaseResponse:
      type: object
      properties:
        workstreamId:
          type: string
        caseId:
          type: string
        controllerAgentRef:
          type: string
        status:
          type: string
        sequence:
          type: integer
          format: int32
        startedAt:
          type: string
        archivedAt:
          type: string
        compactionReason:
          type: string
        contextRevision:
          type: string
      required:
      - caseId
      - controllerAgentRef
      - sequence
      - startedAt
      - status
      - workstreamId
    CompactControllerCaseRequest:
      type: object
      properties:
        caseId:
          type: string
        compactionReason:
          type: string
...
    AttemptSummary:
      type: object
      properties:
        attemptId:
          type: string
        workflowId:
          type: string
        stepId:
          type: string
        agentName:
          type: string
        status:
          type: string
        revision:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
      - agentName
      - attemptId
      - revision
      - status
      - stepId
      - workflowId
    ChangeSummary:
      type: object
      properties:
        kind:
          type: string
        refId:
          type: string
        workflowId:
          type: string
        timestamp:
          type: string
      required:
      - kind
      - refId
      - timestamp
    EnvironmentSection:
      type: object
      properties:
        count:
          type: integer
          format: int32
        byState:
          type: object
          additionalProperties:
            type: integer
            format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/EnvironmentSummary"
        truncated:
          type: boolean
      required:
      - byState
      - count
      - items
      - truncated
    EnvironmentSummary:
      type: object
      properties:
        environmentId:
          type: string
        workflowId:
          type: string
        namespaceId:
          type: string
        lifecycleState:
          type: string
        revision:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
      required:
      - environmentId
      - lifecycleState
      - namespaceId
      - revision
      - workflowId
    HumanActionSummary:
      type: object
      properties:
        interactionId:
          type: string
        workflowId:
          type: string
        stepId:
          type: string
        interactionType:
          type: string
        status:
          type: string
      required:
      - interactionId
      - workflowId
    OracleFailureSummary:
      type: object
      properties:
        executionId:
          type: string
        oracleId:
          type: string
        workflowId:
          type: string
        namespaceId:
          type: string
        updatedAt:
          type: string
      required:
      - executionId
    StepCounts:
      type: object
      properties:
        running:
          type: integer
          format: int32
        waitingHuman:
          type: integer
          format: int32
        blocked:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/StepSummary"
        truncated:
          type: boolean
      required:
      - blocked
      - items
      - running
      - truncated
      - waitingHuman
    StepSummary:
      type: object
      properties:
        workflowId:
          type: string
        stepId:
          type: string
        name:
          type: string
        status:
          type: string
      required:
      - status
      - stepId
      - workflowId
    WorkflowSummary:
      type: object
      properties:
        workflowId:
          type: string
        namespaceId:
          type: string
        workflowType:
          type: string
        title:
          type: string
        status:
          type: string
        revision:
          type: integer
          format: int32
      required:
      - revision
      - workflowId
    WorkstreamProjectionResponse:
      type: object
      properties:
        workstreamId:
          type: string
        namespaceId:
          type: string
        status:
          type: string
        workstreamRevision:
          type: string
        activeWorkflows:
          $ref: "#/components/schemas/WorkstreamSectionWorkflowSummary"
        steps:
          $ref: "#/components/schemas/StepCounts"
        attempts:
          $ref: "#/components/schemas/WorkstreamSectionAttemptSummary"
        humanActions:
          $ref: "#/components/schemas/WorkstreamSectionHumanActionSummary"
        failedOracles:
          $ref: "#/components/schemas/WorkstreamSectionOracleFailureSummary"
        environments:
          $ref: "#/components/schemas/EnvironmentSection"
        recentChanges:
          $ref: "#/components/schemas/WorkstreamSectionChangeSummary"
        boundaryViolations:
          type: integer
          format: int32
      required:
      - activeWorkflows
      - attempts
      - boundaryViolations
      - environments
      - failedOracles
      - humanActions
      - recentChanges
      - status
      - steps
      - workstreamId
      - workstreamRevision
    WorkstreamSectionAttemptSummary:
      type: object
      properties:
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/AttemptSummary"
        truncated:
          type: boolean
      required:
      - count
      - items
      - truncated
    WorkstreamSectionChangeSummary:
      type: object
      properties:
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChangeSummary"
        truncated:
          type: boolean
      required:
      - count
      - items
      - truncated
    WorkstreamSectionHumanActionSummary:
      type: object
      properties:
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/HumanActionSummary"
        truncated:
          type: boolean
      required:
      - count
      - items
      - truncated
    WorkstreamSectionOracleFailureSummary:
      type: object
      properties:
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/OracleFailureSummary"
        truncated:
          type: boolean
      required:
      - count
      - items
      - truncated
    WorkstreamSectionWorkflowSummary:
      type: object
      properties:
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowSummary"
        truncated:
          type: boolean
      required:
      - count
      - items
      - truncated
    ControllerCaseHistoryResponse:
      type: object
      properties:
        workstreamId:
          type: string
        activeCaseId:
          type: string
        cases:
          type: array
          items:
            $ref: "#/components/schemas/ControllerCaseResponse"
      required:
      - cases
      - workstreamId
    ContextBlocker:
      type: object
      properties:
        kind:
          type: string
        refId:
          type: string
        workflowId:
          type: string
      required:
      - kind
      - refId
    ContextChange:
      type: object
      properties:
        kind:
          type: string
        refId:
          type: string
        workflowId:
          type: string
        timestamp:
          type: string
      required:
      - kind
      - refId
      - timestamp
    ContextHumanAction:
      type: object
      properties:
        interactionId:
          type: string
        workflowId:
          type: string
        interactionType:
          type: string
        status:
          type: string
      required:
      - interactionId
      - workflowId
    ContextWorkflow:
      type: object
      properties:
        workflowId:
          type: string
        workflowType:
          type: string
        title:
          type: string
        status:
          type: string
      required:
      - workflowId
    ControllerContextCounts:
      type: object
      properties:
        activeWorkflows:
          type: integer
          format: int32
        running:
          type: integer
          format: int32
        waitingHuman:
          type: integer
          format: int32
        blocked:
          type: integer
          format: int32
        attempts:
          type: integer
          format: int32
        humanActions:
          type: integer
          format: int32
        failedOracles:
          type: integer
          format: int32
        environments:
          type: integer
          format: int32
      required:
      - activeWorkflows
      - attempts
      - blocked
      - environments
      - failedOracles
      - humanActions
      - running
      - waitingHuman
    ControllerResumptionPackage:
      type: object
      properties:
        workstreamId:
          type: string
        sourceRevision:
          type: string
        counts:
          $ref: "#/components/schemas/ControllerContextCounts"
        activeWorkflows:
          type: array
          items:
            $ref: "#/components/schemas/ContextWorkflow"
        openHumanInteractions:
          type: array
          items:
            $ref: "#/components/schemas/ContextHumanAction"
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/ContextBlocker"
        recentChanges:
          type: array
          items:
            $ref: "#/components/schemas/ContextChange"
      required:
      - activeWorkflows
      - blockers
      - counts
      - openHumanInteractions
      - recentChanges
      - sourceRevision
      - workstreamId
...